feat: enhance channel and schedule management; add active channel retrieval, update validation messages, and improve user role handling

+++ qqtech9-backend/src/services/channels.js
@@ -2,12 +2,21 @@ import { Channels } from "../models/channels.js";
 import { ChannelsRepository } from "../repositories/channels.js";
 import { NotFoundError } from "../utils/errors.js";
 import { isValidUuid } from "../utils/validations.js";
+import { AuthService } from "./auth.js";
 
 export const ChannelService = {
-    getAllChannels: async () => { 
+    getAllChannels: async (currentUserFirebaseUid) => {
+        await AuthService.requireAdmin(currentUserFirebaseUid);
+        
         const channels = await ChannelsRepository.findAll();
 
         return Channels.fromArray(channels);
+    },
+
+    getAllChannelsActive: async () => {
+        const channels = await ChannelsRepository.findActiveChannels();
+
+        return channels;
     },
 
     getChannelById: async (id) => {
@@ -24,7 +33,9 @@ export const ChannelService = {
         return channel;
     },
 
-    createChannel: async (dto) => { 
+    createChannel: async (dto, currentUserFirebaseUid) => {
+        await AuthService.requireAdmin(currentUserFirebaseUid);
+
         const newChannel = new Channels(dto);
 
         const savedChannel = await ChannelsRepository.create(newChannel);
@@ -32,7 +43,9 @@ export const ChannelService = {
         return savedChannel;
     },
 
-    updateChannel: async (id, dto) => {
+    updateChannel: async (id, dto, currentUserFirebaseUid) => {
+        await AuthService.requireAdmin(currentUserFirebaseUid);
+
         const existingChannel = await ChannelService.getChannelById(id);
 
         const updatedChannel = new Channels({
@@ -46,7 +59,9 @@ export const ChannelService = {
         return savedChannel;
     },
 
-    deleteChannel: async (id) => {
+    deleteChannel: async (id, currentUserFirebaseUid) => {
+        await AuthService.requireAdmin(currentUserFirebaseUid);
+
         await ChannelService.getChannelById(id);
 
         await ChannelsRepository.delete(id);
